DEL-2348 Transform ActionFilter into middleware
