+++ src/content/Kentico.Kontent.Boilerplate/appsettings.json
@@ -10,5 +10,6 @@
       "Microsoft": "Information"
     }
   },
+  "KenticoKontentWebhookSecret": "",
   "ResponsiveWidths": [ 200, 300, 400, 600, 800, 1000, 1200, 1400, 1600, 2000 ]
 }
